tech-debt: move poll duration consts to own module for react-refresh lint (ROK-1206)
